feat: create a backend transformer
This backend layer is responsible for receiving the items from the connector and 
applying any data transformation necessary that will then be used to output.

This specific backend accepts a connector and apply an item transformer class
to each item and use an aggregate transformer at the end to return one single 
class responsible to handle the items.