feat: add a business layer with a product entity
This layer is where the business logic would live. Everything in this space can 
change fast without impacting the rest of the code.

This class represents a product line. One thing to change here is the usage of the 
quantity, while I was developing I could notice that it makes more sense to let the receipt handle it.