feat: add a receipt entity
This entity is responsible to handle all products and their quantities - or should 
be. It also relies on a tax calculator to perform any tax logic.

This class is currently violating the Single-Responsibility Principle since I'm using
the #to_s method override to actually make a report. A report structure would
fit better and isolate this responsibility.