chore: create a simple runner
This is a way to run the code and see its outputs. Unfortunately, I could not make
it calculate the taxes correctly so far. I focused my time on laying a good 
foundation for the code and there is not time now to fix it nor adding unit tests.
I apologize for that and I hope that you guys appreciate the code structure.